refactor: migrate MogTile to core domain and consolidate tile processing logic into extensions
